Move shared files to shared directory
