+++ backend/src/shared/entities/Edge.ts
@@ -1,0 +1,12 @@
+import { EdgeDescriptor } from './EdgeDescriptor';
+import { Property } from './Property';
+
+/**
+ * Represents an edge of a graph with additional information.
+ */
+export interface Edge extends EdgeDescriptor {
+  type: string;
+  properties: {
+    [key: string]: Property;
+  };
+}
