+++ backend/src/shared/entities/EdgeDescriptor.ts
@@ -1,0 +1,14 @@
+/**
+ * Describes an simple edge of a graph.
+ */
+export interface EdgeDescriptor {
+  id: number;
+  /**
+   * ID of the start node
+   */
+  from: number;
+  /**
+   * ID of the end node
+   */
+  to: number;
+}
