+++ backend/src/shared/entities/NodeDescriptor.ts
@@ -1,0 +1,6 @@
+/**
+ * Describes a simple node of a graph.
+ */
+export interface NodeDescriptor {
+  id: number;
+}
