+++ backend/src/shared/entities/Property.ts
@@ -1,0 +1,5 @@
+/**
+ * Types that are possible for the properties of a neo4j node or relationship
+ * TODO Find out which types are possible
+ */
+export type Property = number | string | Record<string, unknown> | unknown;
