+++ backend/src/shared/queries/LimitQuery.ts
@@ -1,0 +1,6 @@
+export interface LimitQuery {
+  limit?: {
+    nodes?: number;
+    edges?: number;
+  };
+}
