+++ backend/src/shared/queries/QueryResult.ts
@@ -1,0 +1,10 @@
+import { EdgeDescriptor } from '../entities/EdgeDescriptor';
+import { NodeDescriptor } from '../entities/NodeDescriptor';
+
+/**
+ * Generic Result of an API Query
+ */
+export class QueryResult {
+  nodes: NodeDescriptor[];
+  edges: EdgeDescriptor[];
+}
