+++ backend/tests/e2e/app.e2e.spec.ts
@@ -1,14 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AppModule } from '../../src/app.module';
 import { AppService } from '../../src/app.service';
-import { Node } from '../../src/entities/Node';
-import { Edge } from '../../src/entities/Edge';
+import { Node } from '../../src/shared/entities/Node';
+import { Edge } from '../../src/shared/entities/Edge';
 import {
   getEdgesByIdDummies,
   getNodesByIdDummies,
   queryAllDummies,
 } from '../fixtures/testingDumpData';
-import { QueryResult } from '../../src/entities/queries/QueryResult';
+import { QueryResult } from '../../src/shared/queries/QueryResult';
 import { KmapNeo4jModule } from '../../src/config/neo4j/KmapNeo4jModule';
 import { Neo4jService } from 'nest-neo4j/dist';
