+++ shared/src/entities/Edge.ts
