+++ shared/src/entities/EdgeDescriptor.ts
