+++ shared/src/entities/Node.ts
