+++ shared/src/entities/NodeDescriptor.ts
