+++ shared/src/entities/Property.ts
